+++ src/oci/database/models/create_autonomous_database_clone_details.py
@@ -112,6 +112,10 @@ class CreateAutonomousDatabaseCloneDetails(CreateAutonomousDatabaseBase):
             The value to assign to the is_auto_scaling_enabled property of this CreateAutonomousDatabaseCloneDetails.
         :type is_auto_scaling_enabled: bool
 
+        :param is_dev_tier:
+            The value to assign to the is_dev_tier property of this CreateAutonomousDatabaseCloneDetails.
+        :type is_dev_tier: bool
+
         :param is_dedicated:
             The value to assign to the is_dedicated property of this CreateAutonomousDatabaseCloneDetails.
         :type is_dedicated: bool
@@ -257,6 +261,7 @@ class CreateAutonomousDatabaseCloneDetails(CreateAutonomousDatabaseBase):
             'license_model': 'str',
             'is_preview_version_with_service_terms_accepted': 'bool',
             'is_auto_scaling_enabled': 'bool',
+            'is_dev_tier': 'bool',
             'is_dedicated': 'bool',
             'autonomous_container_database_id': 'str',
             'in_memory_percentage': 'int',
@@ -310,6 +315,7 @@ class CreateAutonomousDatabaseCloneDetails(CreateAutonomousDatabaseBase):
             'license_model': 'licenseModel',
             'is_preview_version_with_service_terms_accepted': 'isPreviewVersionWithServiceTermsAccepted',
             'is_auto_scaling_enabled': 'isAutoScalingEnabled',
+            'is_dev_tier': 'isDevTier',
             'is_dedicated': 'isDedicated',
             'autonomous_container_database_id': 'autonomousContainerDatabaseId',
             'in_memory_percentage': 'inMemoryPercentage',
@@ -362,6 +368,7 @@ class CreateAutonomousDatabaseCloneDetails(CreateAutonomousDatabaseBase):
         self._license_model = None
         self._is_preview_version_with_service_terms_accepted = None
         self._is_auto_scaling_enabled = None
+        self._is_dev_tier = None
         self._is_dedicated = None
         self._autonomous_container_database_id = None
         self._in_memory_percentage = None